requirements.txt
flask==3.0.*
tensorflow==2.16.*
numpy==1.26.*
pillow
gunicorn==23.0.*